fixes
-recoded fullscreen

-completely got rid of OpenAL and NAudio, the editor is now using the Bass audio library

-renamed the 'Master' slider to 'Music' and made it only apply to music volume

-fixed offset being limited to 255
-fixed desync
-fixed mouse wheel

+++ Blox Saber Editor/Gui/GuiTextBox.cs
@@ -154,7 +154,7 @@ namespace Blox_Saber_Editor.Gui
 
 			var keyChar = key.ToString();
 			
-			if (Numeric && (!CanBeNegative && !byte.TryParse(keyChar, out _) || CanBeNegative && !sbyte.TryParse(_text.Insert(_cursorPos, keyChar), out _))) //if not a number
+			if (Numeric && (!CanBeNegative && !byte.TryParse(keyChar, out _) || CanBeNegative && !long.TryParse(_text.Insert(_cursorPos, keyChar), out _))) //if not a number
 			{
 				if (Decimal)
 				{
